+++ hdrezka/api/ajax.py
@@ -11,26 +11,26 @@ class AJAX:
     __slots__ = ()
 
     @classmethod
-    def _send_data(cls, action: str, data: dict[str]) -> dict[str]:
-        answer: dict = get_response('POST', f'https://rezka.ag/ajax/{action}/', data=data).json()
+    async def _send_data(cls, action: str, data: dict[str]) -> dict[str]:
+        answer: dict = (await get_response('POST', f'https://rezka.ag/ajax/{action}/', data=data)).json()
         if not answer.get('success', True):
             raise AJAXFail(answer.get('message', answer))
         return answer
 
     @classmethod
-    def get_cdn_series(cls, data: dict[str]):
-        return cls._send_data('get_cdn_series', data)
+    async def get_cdn_series(cls, data: dict[str]):
+        return await cls._send_data('get_cdn_series', data)
 
     @classmethod
-    def get_episodes(cls, id: AnyID, translator_id: AnyID):
-        return cls.get_cdn_series(
+    async def get_episodes(cls, id: AnyID, translator_id: AnyID):
+        return await cls.get_cdn_series(
             {'action': 'get_episodes',
              'id': id,
              'translator_id': translator_id})
 
     @classmethod
-    def get_stream(cls, id: AnyID, translator_id: AnyID, season: AnyID, episode: AnyID):
-        return cls.get_cdn_series(
+    async def get_stream(cls, id: AnyID, translator_id: AnyID, season: AnyID, episode: AnyID):
+        return await cls.get_cdn_series(
             {'action': 'get_stream',
              'id': id,
              'translator_id': translator_id,
@@ -39,8 +39,8 @@ class AJAX:
         )
 
     @classmethod
-    def get_movie(cls, id: AnyID, translator_id: AnyID):
-        return cls.get_cdn_series(
+    async def get_movie(cls, id: AnyID, translator_id: AnyID):
+        return await cls.get_cdn_series(
             {'action': 'get_movie',
              'id': id,
              'translator_id': translator_id}