+++ hdrezka/api/http.py
@@ -1,11 +1,12 @@
-from httpx import request as _request
+import httpx
 
 __all__ = ('get_response', 'DEFAULT_REQUEST_KWARGS')
 
 DEFAULT_REQUEST_KWARGS = {'headers': {'user-agent': 'Mozilla/5.0'}}
 
 
-def get_response(*args, **kwargs):
+async def get_response(*args, **kwargs):
     for k, v in DEFAULT_REQUEST_KWARGS.items():
         kwargs.setdefault(k, v)
-    return _request(*args, **kwargs)
+    async with httpx.AsyncClient() as client:
+        return await client.request(*args, **kwargs)