+++ hdrezka/post/page.py
@@ -1,4 +1,4 @@
-from typing import Iterable, Iterator, Sequence, SupportsIndex, TypeVar
+from typing import Iterable, TypeVar
 
 from ._dataclass import frozen_slots_dataclass
 from .._bs4 import BeautifulSoup
@@ -28,29 +28,19 @@ class InlineItem:
     'Image url'
 
     @property
-    def player(self) -> PlayerType:
+    async def player(self) -> PlayerType:
         """Return a Player Instance"""
-        return Player(self.url)
+        return await Player(self.url)
 
 
 class Page:
     """AJAX class for HDRezka search"""
-    __slots__ = ('_page', '_page_format')
+    __slots__ = ('_page', '_page_format', '__yields', '__yields_page')
 
     def __init__(self, url: str = 'https://rezka.ag/'):
+        self.__yields = []
+        self.__yields_page = 0
         self.page = url
-
-    def __iter__(self) -> Iterator[InlineItem]:
-        """
-        Returns the generator of all found articles
-        """
-        page = 1
-        while True:
-            try:
-                yield from self.page_iter(page)
-            except EmptyPage(page):
-                return
-            page += 1
 
     @property
     def page(self) -> str:
@@ -60,43 +50,45 @@ class Page:
     def page(self, value):
         if not isinstance(value, str):
             value = str(value)
-        if not value.startswith('https://'):
+        # noinspection HttpUrlsUsage
+        if not (value.startswith('https://') or value.startswith('http://')):
             value = f'https://{value}'
         self._page = value
         self._page_format = self._concat_paginator(value.removesuffix('/')).format
 
     @staticmethod
-    def _concat_paginator(url: str):
+    def _concat_paginator(url: str) -> str:
         return f'{url}/page/{{0}}'
 
-    def page_iter(self, page: PageNumber = 1) -> Iterator[InlineItem] | None:
-        page = _range_from_slice(page)
-        is_not_iter = not isinstance(page, Iterable)
-        if is_not_iter:
-            page = page,
-        for page in page:
-            items = BeautifulSoup(get_response('GET', self._page_format(page)).text
-                                  ).find_all(class_='b-content__inline_item')
-            if not items:
-                if is_not_iter:
-                    raise EmptyPage(page)
-                return
-            yield from (InlineItem(
-                (a := (link := i.find(class_='b-content__inline_item-link')).find('a'))['href'],
-                a.text,
-                link.find('div').text,
-                i.find(class_='b-content__inline_item-cover').find('img').get('src', ''))
-                for i in items)
+    def __aiter__(self):
+        self.__yields_page = 0
+        self.__yields.clear()
+        return self
 
-    def get_pages(self, page: PageNumber = 1) -> tuple[InlineItem]:
-        return *self.page_iter(page),
+    async def __anext__(self) -> InlineItem:
+        if not self.__yields:
+            try:
+                self.__yields_page += 1
+                self.__yields += await self.get_page(self.__yields_page)
+            except EmptyPage:
+                raise StopAsyncIteration
+        return self.__yields.pop(0)
 
-    def __getitem__(self, item: PageNumber | Sequence[PageNumber | SupportsIndex | slice]
-                    ) -> tuple[InlineItem] | InlineItem:
-        if isinstance(item, Sequence) and len(item) == 2:
-            page, index = item
-            return self.get_pages(page)[index]
-        return self.get_pages(item)
+    async def get_page(self, page: PageNumber) -> list[InlineItem]:
+        """
+        Get items from pages range.
+        None for all elements of all pages
+        """
+        items = BeautifulSoup((await get_response('GET', self._page_format(page))).text
+                              ).find_all(class_='b-content__inline_item')
+        if not items:
+            raise EmptyPage(page)
+        return [InlineItem(
+            (a := (link := i.find(class_='b-content__inline_item-link')).find('a'))['href'],
+            a.text,
+            link.find('div').text,
+            i.find(class_='b-content__inline_item-cover').find('img').get('src', ''))
+            for i in items]
 
     def __repr__(self):
         return f'{self.__class__.__qualname__}({self.page!r})'