+++ hdrezka/stream/player.py
@@ -1,5 +1,5 @@
+import sys
 from collections import defaultdict
-from functools import lru_cache
 from typing import Any, SupportsInt, TypeVar
 
 from .._bs4 import BeautifulSoup
@@ -10,16 +10,41 @@ from ..post import *
 __all__ = ('Player', 'PlayerType', 'PlayerBase', 'PlayerMovie', 'PlayerSeries')
 
 
+class _CacheStorage:
+    __slots__ = ('max_size', '__storage')
+
+    def __init__(self, max_size: int = 256):
+        self.max_size = max_size
+        self.__storage = {}
+
+    def get(self, key):
+        return self.__storage.get(key)
+
+    def set(self, key, value):
+        self.__storage[key] = value
+        for k in self.__storage.keys():
+            if sys.getsizeof(self.__storage) <= self.max_size:
+                break
+            del self.__storage[k]
+
+
+__CACHE = _CacheStorage()
+
+
 class PlayerBase:
     __slots__ = ('post',)
 
     def __init__(self, url_or_cast: Any):
+        """Also you should call the `ainit` method"""
         if isinstance(url_or_cast, PlayerBase):
             self.post = url_or_cast.post
             return
         elif not isinstance(url_or_cast, str):
             url_or_cast = str(url_or_cast)
         self.post = Post(url_or_cast)
+
+    async def ainit(self):
+        await self.post.ainit()
 
     def _translator(self, translator_id: SupportsInt = None) -> int:
         if translator_id is None:
@@ -34,40 +59,47 @@ class PlayerBase:
 class PlayerMovie(PlayerBase):
     __slots__ = ()
 
-    def get_stream(self, translator_id: SupportsInt = None) -> URLs:
-        return urls_from_ajax_response(AJAX.get_movie(self.post.id, self._translator(translator_id)))
+    async def get_stream(self, translator_id: SupportsInt = None) -> URLs:
+        return urls_from_ajax_response(await AJAX.get_movie(self.post.id, self._translator(translator_id)))
 
 
 class PlayerSeries(PlayerBase):
     __slots__ = ()
 
-    def get_episodes(self, translator_id: SupportsInt = None) -> defaultdict[int, tuple[int]]:
-        episodes = BeautifulSoup(AJAX.get_episodes(self.post.id, self._translator(translator_id))['episodes'])
+    async def get_episodes(self, translator_id: SupportsInt = None) -> defaultdict[int, tuple[int]]:
+        episodes = BeautifulSoup((await AJAX.get_episodes(self.post.id, self._translator(translator_id)))['episodes'])
         result: defaultdict[int, tuple[int]] = defaultdict(tuple)
         for i in episodes.find_all(class_='b-simple_episode__item', attrs=('data-season_id', 'data-episode_id')):
             result[int(i.get('data-season_id'))] += int(i.get('data-episode_id')),
         return result
 
-    def get_stream(self, season: int, episode: int, translator_id: SupportsInt = None) -> URLs:
-        return urls_from_ajax_response(AJAX.get_stream(self.post.id, self._translator(translator_id), season, episode))
+    async def get_stream(self, season: int, episode: int, translator_id: SupportsInt = None) -> URLs:
+        return urls_from_ajax_response(
+            await AJAX.get_stream(self.post.id, self._translator(translator_id), season, episode))
 
 
 PlayerType = TypeVar('PlayerType', PlayerBase, PlayerMovie, PlayerSeries)
 
 
-@lru_cache(512)
-def player(url_or_path: Any) -> PlayerType:
+async def player(url_or_path: Any) -> PlayerType:
     """
     Returns either Player Series if series, or PlayerMovie if movie, otherwise raises UnknownContentType
     """
     cast = PlayerBase(url_or_path)
+    cached = __CACHE.get(cast.post.url)
+    if cached is not None:
+        return cached
+    await cast.ainit()
     type = cast.post.type
     match type:
         case 'tv_series':
-            return PlayerSeries(cast)
+            value = PlayerSeries(cast)
         case 'movie':
-            return PlayerMovie(cast)
-    raise UnknownContentType(type)
+            value = PlayerMovie(cast)
+        case _:
+            raise UnknownContentType(type)
+    __CACHE.set(cast.post.url, value)
+    return value
 
 
 Player = player